修改“删除非热门博客异常”的bug

+++ src/main/java/com/zlw/blog/service/impl/HotBlogServiceImpl.java
@@ -42,6 +42,8 @@ public class HotBlogServiceImpl implements HotBlogService {
     @Override
     public void delBlogById(Integer blogId) {
         HotBlog hotBlog = hotblogRepository.findByBlogId(blogId);
-        hotblogRepository.delete(hotBlog);
+        if(hotBlog!=null){
+            hotblogRepository.delete(hotBlog);
+        }
     }
 }
